requirements.txt
annotated-doc==0.0.4
python-multipart>=0.0.9
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
torchvision==0.18.1+cpu
python-doctr==1.0.1